+++ app/base_decorator.rb
@@ -1,0 +1,12 @@
+
+require_relative 'Nameable';
+
+class Decorator < Nameable
+  def initialize(nameable)
+    @nameable = nameable
+  end
+
+  def correct_name
+    @nameable.correct_name
+  end
+end
